gpu-service/processors/cloth_swap.py
import torch
import cv2
import numpy as np
from PIL import Image
from typing import Dict, Any
import logging

logger = logging.getLogger(__name__)

class ClothSwapProcessor:
    """Virtual try-on / Cloth swap engine"""
    
    def __init__(self, gpu_manager):
        self.gpu_manager = gpu_manager
        self.model_name = "cloth-swap-model"
    
    def load_model(self):
        """Load cloth swap model (placeholder for actual model)"""
        # In production, load actual virtual try-on model like:
        # - OOTDiffusion
        # - IDM-VTON
        # - Outfit Anyone
        
        logger.info("Loading cloth swap model...")
        # Placeholder - implement actual model loading
        return {"model": "cloth_swap_v1"}
    
    def preprocess_images(self, person_path: str, cloth_path: str):
        """Preprocess person and cloth images"""
        person_img = Image.open(person_path).convert('RGB')
        cloth_img = Image.open(cloth_path).convert('RGB')
        
        # Resize to standard size
        person_img = person_img.resize((768, 1024))
        cloth_img = cloth_img.resize((768, 1024))
        
        return person_img, cloth_img
    
    def detect_pose(self, image: Image.Image):
        """Detect human pose keypoints"""
        # Use OpenPose or MediaPipe for pose detection
        import mediapipe as mp
        
        mp_pose = mp.solutions.pose
        pose = mp_pose.Pose(static_image_mode=True)
        
        image_np = np.array(image)
        results = pose.process(cv2.cvtColor(image_np, cv2.COLOR_RGB2BGR))
        
        return results.pose_landmarks if results else None
    
    def segment_cloth(self, image: Image.Image):
        """Segment cloth region from image"""
        # Use segmentation model to extract cloth
        # Placeholder implementation
        return image
    
    def process(self, job_data: Dict[str, Any]) -> Dict[str, Any]:
        """Perform cloth swap"""
        person_url = job_data.get('personUrl')
        cloth_url = job_data.get('clothUrl')
        category = job_data.get('category', 'upper_body')
        preserve_face = job_data.get('preserveFace', True)
        
        logger.info(f"Processing cloth swap: {category}")
        
        # Download images (implement actual download)
        person_path = f"/tmp/person_{job_data['jobId']}.jpg"
        cloth_path = f"/tmp/cloth_{job_data['jobId']}.jpg"
        
        # Preprocess
        person_img, cloth_img = self.preprocess_images(person_path, cloth_path)
        
        # Detect pose
        pose_landmarks = self.detect_pose(person_img)
        
        with self.gpu_manager.model_context('cloth-swap', self.load_model, required_vram_mb=3000):
            # Perform cloth swap (implement actual inference)
            # This is a placeholder - integrate actual model
            
            result_img = person_img  # Placeholder
            
            # Save result
            output_path = f"/tmp/cloth_swap_{job_data['jobId']}.png"
            result_img.save(output_path)
        
        return {
            'output_image': output_path,
            'category': category
        }
