gpu-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
torch==2.1.2
torchvision==0.16.2
diffusers==0.25.1
transformers==4.37.2
accelerate==0.26.1
xformers==0.0.23.post1
opencv-python==4.9.0.80
pillow==10.2.0
numpy==1.26.3
redis==5.0.1
pydantic==2.5.3
python-multipart==0.0.6
aiofiles==23.2.1
safetensors==0.4.1
compel==2.0.2
controlnet-aux==0.0.7
mediapipe==0.10.9
trimesh==4.0.10
pyrender==0.1.45
moviepy==1.0.3
pydub==0.25.1
TTS==0.22.0
